+++ .github/CHANGELOG.md
@@ -10,6 +10,9 @@
 * Added a ``rich`` based repr to ``ParameterSet``.
 [(#616)](https://github.com/XanaduAI/MrMustard/pull/616)
 
+* Made `JAX` an optional `jax_backend` dependency.
+[(#637)](https://github.com/XanaduAI/MrMustard/pull/637)
+
 * Added vjps to the JAX backend. Added ``numba`` based Fock lattice strategies to `SqueezedVacuum` and `Sgate`.
 [(#636)](https://github.com/XanaduAI/MrMustard/pull/636)